providers: reorganize to avoid needless code repetition
