+++ ibp/providers/tdcj.py
@@ -1,17 +1,18 @@
 """TDCJ inmate query implementation."""
 
-import asyncio
 import datetime
 import logging
+import re
 import typing
 from urllib.parse import urljoin
 
 from bs4 import BeautifulSoup, Tag
 from nameparser import HumanName
 
-from .decorators import log_query_by_inmate_id, log_query_by_name
+from .misc import run_curl_exec
+from .types import QueryResult
 
-LOGGER = logging.getLogger(__name__)
+LOGGER: logging.Logger = logging.getLogger(__name__)
 
 BASE_URL = "https://inmate.tdcj.texas.gov"
 SEARCH_PATH = "InmateSearch/search.action"
@@ -24,34 +25,13 @@ def format_inmate_id(inmate_id: typing.Union[int, str]) -> str:
     return f"{inmate_id:08d}"
 
 
-class QueryResult(typing.TypedDict):
-    """Result of a TDCJ query."""
-
-    id: str
-    jurisdiction: typing.Literal["Texas"]
-
-    first_name: str
-    last_name: str
-
-    unit: str
-
-    race: typing.Optional[str]
-    sex: typing.Optional[str]
-
-    url: str
-    release: typing.Optional[str | datetime.date]
-
-    datetime_fetched: datetime.datetime
-
-
 async def _curl_search_url(
     last_name: str = "",
     first_name: str = "",
     inmate_id: str = "",
     timeout: float | None = None,
 ):
-    command = [
-        "curl",
+    args = [
         "--ipv4",
         "-d",
         "btnSearch=Search",
@@ -72,30 +52,10 @@ async def _curl_search_url(
         SEARCH_URL,
     ]
 
-    if timeout is not None:
-        command.extend(["--max-time", str(float(timeout))])
-
-    process = await asyncio.create_subprocess_exec(
-        *command, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
-    )
-
-    try:
-        stdout, *_ = await asyncio.wait_for(process.communicate(), timeout=timeout)
-    except asyncio.TimeoutError as error:
-        if process.returncode is None:  # Check if process is still running
-            process.terminate()
-            await process.wait()  # Wait for the process to actually terminate
-        message = f"curl command timed out after {timeout} seconds."
-        raise TimeoutError(message) from error
-
-    if process.returncode != 0:
-        message = f"curl command failed with exit code {process.returncode}"
-        raise RuntimeError(message)
-
-    return stdout
+    return await run_curl_exec(args, timeout=timeout)
 
 
-async def _query(  # pylint: disable=too-many-locals
+async def query(  # pylint: disable=too-many-locals
     last_name: str = "",
     first_name: str = "",
     inmate_id: str = "",
@@ -138,6 +98,11 @@ async def _query(  # pylint: disable=too-many-locals
     def entry_to_inmate(entry: dict):
         """Convert TDCJ inmate entry to inmate dictionary."""
 
+        def parse_inmate_id(inmate_id: str) -> int:
+            return int(re.sub(r"\D", "", inmate_id))
+
+        inmate_id = parse_inmate_id(entry["inmateNum"])
+
         name = HumanName(entry.get("Name", ""))
 
         def build_url(href):
@@ -156,7 +121,7 @@ async def _query(  # pylint: disable=too-many-locals
             LOGGER.debug("Failed to parse release date '%s'", release)
 
         return QueryResult(
-            id=entry["TDCJ Number"],
+            id=inmate_id,
             jurisdiction="Texas",
             first_name=name.first,
             last_name=name.last,
@@ -169,21 +134,3 @@ async def _query(  # pylint: disable=too-many-locals
         )
 
     return list(map(entry_to_inmate, entries))
-
-
-@log_query_by_name(LOGGER)
-async def query_by_name(first, last, **kwargs):
-    """Query the TDCJ database with an inmate name."""
-    return await _query(first_name=first, last_name=last, **kwargs)
-
-
-@log_query_by_inmate_id(LOGGER)
-async def query_by_inmate_id(inmate_id: str | int, **kwargs):
-    """Query the TDCJ database with an inmate id."""
-    try:
-        inmate_id = format_inmate_id(inmate_id)
-    except ValueError as exc:
-        msg = f"'{inmate_id}' is not a valid Texas inmate number"
-        raise ValueError(msg) from exc
-
-    return await _query(inmate_id=inmate_id, **kwargs)
